+++ packages/bitswap/src/block-broker.ts
@@ -1,9 +1,8 @@
-import { createBitswap } from '@helia/bitswap'
-import { isPeerId, start } from '@libp2p/interface'
+import { isPeerId } from '@libp2p/interface'
 import { CustomProgressEvent } from 'progress-events'
-import type { BitswapOptions, Bitswap, BitswapWantBlockProgressEvents, BitswapNotifyProgressEvents } from '@helia/bitswap'
-import type { BlockAnnounceOptions, BlockBroker, BlockRetrievalOptions, CreateSessionOptions, Routing, HasherLoader, SessionBlockBroker, BlockBrokerConnectProgressEvent, BlockBrokerConnectedProgressEvent, BlockBrokerRequestBlockProgressEvent, BlockBrokerReceiveBlockProgressEvent, HeliaMixin } from '@helia/interface'
-import type { HeliaWithLibp2p } from '@helia/libp2p'
+import { Bitswap } from './bitswap.ts'
+import type { BitswapOptions, BitswapWantBlockProgressEvents, BitswapNotifyProgressEvents } from './index.ts'
+import type { BlockAnnounceOptions, BlockBroker, BlockRetrievalOptions, CreateSessionOptions, Routing, HasherLoader, SessionBlockBroker, BlockBrokerConnectProgressEvent, BlockBrokerConnectedProgressEvent, BlockBrokerRequestBlockProgressEvent, BlockBrokerReceiveBlockProgressEvent } from '@helia/interface'
 import type { Libp2p, Startable, ComponentLogger } from '@libp2p/interface'
 import type { Blockstore } from 'interface-blockstore'
 import type { CID } from 'multiformats/cid'
@@ -20,13 +19,13 @@ export interface BitswapBlockBrokerInit extends BitswapOptions {
 
 }
 
-class BitswapBlockBroker implements BlockBroker<BitswapWantBlockProgressEvents, BitswapNotifyProgressEvents>, Startable {
+export class BitswapBlockBroker implements BlockBroker<BitswapWantBlockProgressEvents, BitswapNotifyProgressEvents>, Startable {
   public readonly name = 'bitswap'
   private readonly bitswap: Bitswap
   private started: boolean
 
   constructor (components: BitswapBlockBrokerComponents, init: BitswapBlockBrokerInit = {}) {
-    this.bitswap = createBitswap(components, init)
+    this.bitswap = new Bitswap(components, init)
     this.started = false
   }
 
@@ -117,32 +116,4 @@ class BitswapBlockBroker implements BlockBroker<BitswapWantBlockProgressEvents, 
       }
     }
   }
-}
-
-/**
- * A helper factory for users who want to override Helia `blockBrokers` but
- * still want to use the default `BitswapBlockBroker`.
- */
-export function bitswap (init: BitswapBlockBrokerInit = {}): (components: BitswapBlockBrokerComponents) => BlockBroker {
-  return (components) => new BitswapBlockBroker(components, init)
-}
-
-/**
- * Return a Helia node augmented with a libp2p instance
- */
-export function withBitswap (helia: HeliaWithLibp2p, options: BitswapBlockBrokerInit = {}): HeliaWithLibp2p {
-  const mixin: HeliaMixin<HeliaWithLibp2p> = {
-    start: async (helia) => {
-      if (helia.status === 'starting' && !helia.hasBlockBroker('bitswap')) {
-        const broker = new BitswapBlockBroker(helia, options)
-        await start(broker)
-
-        helia.addBlockBroker(broker)
-      }
-    }
-  }
-
-  helia.addMixin(mixin)
-
-  return helia
 }
