+++ packages/block-brokers/package.json
@@ -58,7 +58,6 @@
     "@multiformats/multiaddr-matcher": "^3.0.2",
     "@multiformats/multiaddr-to-uri": "^12.0.0",
     "@multiformats/uri-to-multiaddr": "^10.0.0",
-    "interface-blockstore": "^7.0.1",
     "multiformats": "^14.0.0",
     "progress-events": "^1.1.0",
     "uint8arraylist": "^3.0.2",
