+++ packages/http/package.json
@@ -49,23 +49,13 @@
   },
   "dependencies": {
     "@helia/block-brokers": "^5.2.4",
-    "@helia/delegated-routing-v1-http-api-client": "^7.0.1",
     "@helia/interface": "^6.2.1",
-    "@helia/routers": "^5.1.1",
-    "@helia/utils": "^2.5.2",
-    "@libp2p/config": "^1.1.27",
-    "@libp2p/interface": "^3.2.0",
-    "@libp2p/keychain": "^6.0.12",
-    "@multiformats/dns": "^1.0.13",
-    "blockstore-core": "^7.0.1",
-    "datastore-core": "^12.0.1",
-    "interface-datastore": "^10.0.1",
-    "libp2p": "^3.2.0"
+    "@helia/routers": "^5.1.1"
   },
   "devDependencies": {
+    "@libp2p/interface": "^3.2.3",
     "aegir": "^48.0.4",
     "helia": "^6.1.4",
-    "multiformats": "^14.0.0",
     "sinon": "^22.0.0",
     "sinon-ts": "^2.0.0"
   },
