+++ packages/interface/package.json
@@ -74,6 +74,7 @@
   "dependencies": {
     "@ipshipyard/crypto": "^1.1.0",
     "@ipshipyard/keychain": "^1.0.2",
+    "@libp2p/interface": "^3.2.3",
     "@multiformats/dns": "^1.0.13",
     "@multiformats/multiaddr": "^13.0.1",
     "abort-error": "^1.0.2",
