+++ packages/interop/package.json
@@ -57,8 +57,8 @@
     "@helia/dag-cbor": "^5.1.1",
     "@helia/dag-json": "^5.1.1",
     "@helia/dnslink": "^1.2.1",
-    "@helia/http": "^3.1.4",
     "@helia/interface": "^6.2.1",
+    "@helia/libp2p": "^0.0.0",
     "@helia/ipns": "^9.2.1",
     "@helia/json": "^5.1.1",
     "@helia/mfs": "^7.1.1",
