+++ packages/libp2p/package.json
@@ -48,21 +48,44 @@
     "test:electron-main": "aegir test -t electron-main"
   },
   "dependencies": {
-    "@helia/block-brokers": "^5.2.4",
+    "@chainsafe/libp2p-noise": "^17.0.0",
+    "@chainsafe/libp2p-yamux": "^8.0.1",
     "@helia/delegated-routing-v1-http-api-client": "^7.0.1",
     "@helia/interface": "^6.2.1",
     "@helia/routers": "^5.1.1",
-    "@helia/utils": "^2.5.2",
+    "@ipshipyard/libp2p-auto-tls": "^2.0.1",
+    "@libp2p/autonat": "^3.0.21",
+    "@libp2p/bootstrap": "^12.0.24",
+    "@libp2p/circuit-relay-v2": "^4.2.6",
+    "@libp2p/config": "^1.1.32",
+    "@libp2p/dcutr": "^3.0.21",
+    "@libp2p/http": "^2.0.2",
+    "@libp2p/identify": "^4.1.7",
+    "@libp2p/interface": "^3.2.3",
+    "@libp2p/kad-dht": "^16.3.2",
+    "@libp2p/keychain": "^6.1.2",
+    "@libp2p/mdns": "^12.0.24",
+    "@libp2p/mplex": "^12.0.24",
+    "@libp2p/peer-id": "^6.0.10",
+    "@libp2p/ping": "^3.1.6",
+    "@libp2p/tcp": "^11.0.21",
+    "@libp2p/tls": "^3.1.3",
+    "@libp2p/upnp-nat": "^4.0.21",
+    "@libp2p/webrtc": "^6.0.24",
+    "@libp2p/websockets": "^10.1.14",
     "@multiformats/dns": "^1.0.13",
-    "blockstore-core": "^7.0.1",
-    "datastore-core": "^12.0.1",
-    "interface-datastore": "^10.0.1"
+    "interface-datastore": "^10.0.1",
+    "it-map": "^3.1.6",
+    "libp2p": "^3.3.3",
+    "multiformats": "^14.0.0"
   },
   "devDependencies": {
+    "@multiformats/multiaddr": "^13.0.3",
     "aegir": "^48.0.11",
+    "datastore-core": "^12.0.1",
     "helia": "^6.1.4",
-    "multiformats": "^14.0.0",
-    "sinon": "^22.0.0",
+    "it-all": "^3.0.11",
+    "it-drain": "^3.0.12",
     "sinon-ts": "^2.0.0"
   },
   "browser": {
