+++ packages/libp2p/test/routing.spec.ts
@@ -3,7 +3,7 @@ import { expect } from 'aegir/chai'
 import drain from 'it-drain'
 import { CID } from 'multiformats'
 import { stubInterface } from 'sinon-ts'
-import { libp2pRouting } from '../src/index.ts'
+import { libp2pRouting } from '../src/routing.ts'
 import type { Router } from '@helia/interface'
 import type { ContentRouting, Libp2p, PeerRouting } from '@libp2p/interface'
 import type { StubbedInstance } from 'sinon-ts'
