+++ packages/utils/package.json
@@ -58,9 +58,12 @@
     "@libp2p/utils": "^7.0.15",
     "@multiformats/dns": "^1.0.13",
     "@multiformats/multiaddr": "^13.0.1",
+    "abort-error": "^1.0.2",
     "any-signal": "^4.2.0",
+    "birnam": "^1.0.0",
     "blockstore-core": "^7.0.1",
     "cborg": "^5.1.0",
+    "datastore-core": "^12.0.1",
     "interface-blockstore": "^7.0.1",
     "interface-datastore": "^10.0.1",
     "it-drain": "^3.0.12",
@@ -68,7 +71,6 @@
     "it-foreach": "^2.1.6",
     "it-merge": "^3.0.13",
     "it-to-buffer": "^5.0.0",
-    "libp2p": "^3.2.0",
     "mortice": "^3.3.1",
     "multiformats": "^14.0.0",
     "p-defer": "^4.0.1",
@@ -81,9 +83,7 @@
     "@libp2p/logger": "^6.2.4",
     "@libp2p/peer-id": "^6.0.6",
     "@types/sinon": "^21.0.1",
-    "abort-error": "^1.0.2",
     "aegir": "^48.0.4",
-    "datastore-core": "^12.0.1",
     "delay": "^7.0.0",
     "it-all": "^3.0.11",
     "it-map": "^3.1.5",
